requirements.txt
altair==5.4.1
asttokens==3.0.1
attrs==25.3.0
backcall==0.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.42.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.6
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.52
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
pywin32==311
pyzmq==27.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rpds-py==0.20.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.40.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.6.0
zipp==3.20.2
